# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GymLink API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List gym members
      parameters:
        - name: gym
          in: query
          schema:
            type: string
        - name: interest
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user's profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteMe
      tags: [users]
      summary: Permanently delete the current user's account
      responses:
        "204":
          description: Account deleted
  /users/me/checkin:
    post:
      operationId: checkIn
      tags: [users]
      summary: Check in or out of a gym
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInBody"
      responses:
        "200":
          description: Updated user profile with check-in status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
...
  /videos:
    get:
      operationId: listWorkoutVideos
      tags: [videos]
      summary: List workout form videos for a user
      parameters:
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkoutVideo"
    post:
      operationId: createWorkoutVideo
      tags: [videos]
      summary: Upload a workout form video
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkoutVideoBody"
      responses:
        "201":
          description: Video created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutVideo"
  /videos/{id}:
    patch:
      operationId: updateWorkoutVideo
      tags: [videos]
      summary: Update a workout video title
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkoutVideoBody"
      responses:
        "200":
          description: Updated video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutVideo"
    delete:
      operationId: deleteWorkoutVideo
      tags: [videos]
      summary: Delete a workout video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /videos/{id}/like:
    post:
      operationId: toggleVideoLike
      tags: [videos]
      summary: Toggle like on a workout video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated like status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoLikeStatus"
  /videos/{id}/likes:
    get:
      operationId: getVideoLikes
      tags: [videos]
      summary: Get like count and whether the current user liked a video
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Like info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoLikeStatus"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        age:
          type: integer
        avatar:
          type: string
          description: Emoji or image URL
        bio:
          type: string
        gym:
          type: string
        gymId:
          type: string
          nullable: true
        schedule:
          type: string
        interests:
          type: array
          items:
            type: string
        verified:
          type: boolean
        distance:
          type: string
        isMe:
          type: boolean
        activeNow:
          type: boolean
        checkedIn:
          type: boolean
        avatarUrl:
          type: string
          nullable: true
        hidden:
          type: boolean
      required:
        - id
        - name
        - age
        - avatar
        - bio
        - gym
        - schedule
        - interests
        - verified
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
        bio:
          type: string
        gym:
          type: string
        gymId:
          type: string
        schedule:
          type: string
        interests:
          type: array
          items:
            type: string
        avatarUrl:
          type: string
          nullable: true
        hidden:
          type: boolean
    CheckInBody:
      type: object
      properties:
        gymId:
          type: string
        gymName:
          type: string
      required:
        - gymId
        - gymName
...
    WorkoutVideo:
      type: object
      required: [id, userId, objectPath, title, createdAt, likeCount, likedByMe]
      properties:
        id:
          type: string
        userId:
          type: string
        objectPath:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        likeCount:
          type: integer
        likedByMe:
          type: boolean
    CreateWorkoutVideoBody:
      type: object
      required: [objectPath, title]
      properties:
        objectPath:
          type: string
        title:
          type: string
        description:
          type: string
    UpdateWorkoutVideoBody:
      type: object
      required: [title]
      properties:
        title:
          type: string
    VideoLikeStatus:
      type: object
      required: [videoId, likeCount, likedByMe]
      properties:
        videoId:
          type: string
        likeCount:
          type: integer
        likedByMe:
          type: boolean